Fixed memory leak in DataValueReference class
Due to the "DataModifiedEvent" being used in the DataValueReference
class, they could never be deleted. And these objects are copied a lot,
so in practice this happened a lot.

One particularly weird thing that happens is that the event handler
fires only on the "original" object, and not the one that's actually in
use. Events weren't being copied properly in the copy constructor.

I don't know if this caused any practical problems. In any case, it's
using the new "WeakEventWrapper" class to allow it to be freed, and the
events are now copied over in the copy constructor.

It's still possible that events may fire twice if the garbage collector
doesn't free up the original object first...