Fixed random crashes
This appears to have been caused by the Cairo.ImageSurface constructor
which took an array of pixel data. Apparently this array is expected to
exist for the lifetime of the Surface, which obviously means it can't be
a managed object.

This crash was triggerable by scrolling through the tileset cloner's
preview tilesets quickly, presumably because this is generating and
destroying hundreds of images each time a new preview is generated.

See: https://github.com/GtkSharp/GtkSharp/issues/442