Undoable ConstantsMappings
This turned out to be necessary to fix a crash with TreasureObjects, as
they add entries to TreasureObjectMappings. This fixes a crash that was
causable with the following steps:

- Create a new treasure object
- Undo
- Toggle the chest's subID value to something else, then back to the
  deleted value

This caused a crash because it detected the name of the newly
created/uncreated treasure object ID in the ConstantsMapping and tried
to use that, but that definition no longer existed in the project's
state.