UI: Replace imgui exception handler with SDL messagebox
This is more resilient, as I don't need to rely on my whole graphics
stack being functional for the exception handler to work.

In fact, I think since updating ImGui.NET, I've been unable to bring
ImGui back to a working condition when an exception skips over
ImGui.End() calls. So an alternative was needed.