LynnaLab: Replace LazyInvoke with DoNextFrame
They were both very similar, the main difference was that LazyInvoke
used to be used for tileset rendering. That's now irrelevant.

The other difference was that DoNextFrame was called after imgui
rendering was done, which is less problematic for a lot of things, ie.
stuff that might cause rendering surfaces to get deleted. In particular,
this fixes a bug where multiple clients accessing the dungeon editor
could crash if one deleted a floor the other was already on.